feat: Get Confirmation of Deletion

+++ src/main/webapp/view.jsp
@@ -114,7 +114,7 @@
 				if (userID != null && userID.equals(board.getUserID())) {
 			%>
 				<a href="update.jsp?boardID=<%= boardID%>" class="btn btn-primary">수정</a>
-				<a href="deleteAction.jsp?boardID=<%= boardID%>" class="btn btn-primary">삭제</a>
+				<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?boardID=<%= boardID%>" class="btn btn-primary">삭제</a>
 			<%
 				}
 			%>
